Update wikipedia.py
If you tried to search for 'ireland' with the auto_suggest=True, the title was set to 'island' because the suggestion='island' and results[0]='ireland'. Thus suggestion or results[0] = suggestion. Shouldn't the expected behavior title be 'ireland' if you searched for 'ireland', even with auto_suggest=True. A small change but make a difference in my opinion.

P.S. reversing the order doesn't work because of you made a spelling mistake, for instance 'ieland', the title will be set to something completely different, 'Jimmy MacCarthy' in that case. I that case you what the suggestion as the title, as in that example the suggestion is actually 'ireland'.